dcs: Drop the separate lbs claimer, update the ttgc claimer
