+++ CHANGELOG.md
@@ -12,7 +12,7 @@ For details about compatibility between different releases, see the **Commitment
 ### Added
 
 - Add tracing for LBS LNS and TTIGW protocol handlers.
-- TTGC LBS Root CUPS claimer.
+- TTGC LBS Root CUPS claiming support.
 
 ### Changed
