+++ pkg/deviceclaimingserver/gateways/gateways.go
@@ -23,7 +23,6 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/cluster"
 	"go.thethings.network/lorawan-stack/v3/pkg/config"
 	"go.thethings.network/lorawan-stack/v3/pkg/config/tlsconfig"
-	"go.thethings.network/lorawan-stack/v3/pkg/deviceclaimingserver/gateways/lbscups"
 	"go.thethings.network/lorawan-stack/v3/pkg/deviceclaimingserver/gateways/ttgc"
 	dcstypes "go.thethings.network/lorawan-stack/v3/pkg/deviceclaimingserver/types"
 	"go.thethings.network/lorawan-stack/v3/pkg/errors"
@@ -49,9 +48,8 @@ type Config struct {
 }
 
 var (
-	errInvalidUpstream   = errors.DefineInvalidArgument("invalid_upstream", "upstream `{name}` is invalid")
-	errTTGCNotEnabled    = errors.DefineFailedPrecondition("ttgc_not_enabled", "TTGC is not enabled")
-	errLBSCUPSNotEnabled = errors.DefineFailedPrecondition("lbs_cups_not_enabled", "TTGC LBS CUPS is not enabled")
+	errInvalidUpstream = errors.DefineInvalidArgument("invalid_upstream", "upstream `{name}` is invalid")
+	errTTGCNotEnabled  = errors.DefineFailedPrecondition("ttgc_not_enabled", "TTGC is not enabled")
 )
 
 // ParseGatewayEUIRanges parses the configured upstream map and returns map of ranges.
@@ -141,13 +139,6 @@ func NewUpstream(
 		}
 		hosts["ttgc"] = ttgcRanges
 	}
-	if _, lbscupsAdded := hosts["lbs-cups"]; ttgcConf.LBSCUPSEnabled && !lbscupsAdded {
-		lbscupsRanges := make([]dcstypes.EUI64Range, len(ttgcConf.LBSCUPSGatewayEUIs))
-		for i, prefix := range ttgcConf.LBSCUPSGatewayEUIs {
-			lbscupsRanges[i] = dcstypes.RangeFromEUI64Prefix(prefix)
-		}
-		hosts["lbs-cups"] = lbscupsRanges
-	}
 
 	// Setup upstream table.
 	for name, ranges := range hosts {
@@ -161,14 +152,6 @@ func NewUpstream(
 				return nil, errTTGCNotEnabled.New()
 			}
 			claimer, err = ttgc.New(ctx, c, ttgcConf)
-			if err != nil {
-				return nil, err
-			}
-		case "lbs-cups":
-			if !ttgcConf.LBSCUPSEnabled {
-				return nil, errLBSCUPSNotEnabled.New()
-			}
-			claimer, err = lbscups.New(ctx, c, ttgcConf)
 			if err != nil {
 				return nil, err
 			}
